test_cc: drive the battery through the warm image (15x), silence the short-overflow warning
The gcc-differential battery and every flag/link scenario booted `ai aicc`
from the source script each time -- ~0.68s of compiler cold-start per compile,
~100 of them. Route them through the WARM-baked aicc.image instead (what `make
install` ships and users run): ~0.02s cold, so `make test_cc` drops 112s -> 7.5s.
The image is added as a prereq, so it re-bakes when any crew/cc/*.l changes; the
laws leg still feeds cc source through plain `$m` for the lex/parse/gen goldens.

23-short.c narrowed 70000 into a short implicitly, tripping gcc -Woverflow on
the differential compile. Make the truncation explicit with a (short) cast --
same intent, same 49, no warning even under -Wall.

+++ test/cc/23-short.c
@@ -1,1 +1,1 @@
-int main() { short s = 70000; long r = s / 100; return r + 5; }
+int main() { short s = (short)70000; long r = s / 100; return r + 5; }
